requirements.txt
Django==4.2
django-ckeditor==6.7.0
djangorestframework==3.14.0
google-cloud-translate==2.0.1
redis==4.6.0
python-dotenv==1.0.0